Split example.go into multiple folders
This creates one folder per component.
